// client/model/networkObjectGroup.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NetworkObject } from './networkObject';


export interface NetworkObjectGroup { 
    readonly id?: string;
    readonly createdAt?: string;
    readonly updatedAt?: string;
    readonly version?: number;
    readonly tenantVersion?: number;
    readonly deletedAt?: string;
    readonly provisionedAt?: string;
    readonly provisionedVersion?: number;
    name: string;
    description?: string;
    readonly networkObjects?: Array<NetworkObject>;
    tierId: string;
}